Add files via upload
Version 4.1, some features work, many don't about to refactor the project to make it easier to work with now that its grown to this point. Overall, i'm happy with the shape and direction this is going. Mainly it seems I have some data analysis bugs to squash, and my annotation system is broken af, but one step at a time this will become a really useful tool for quickly analyzing data.